Poprawki utils: Generyczne drzewo - przechodzenie, linearyzacja kinematiclib - Szkielet na bazie generycznego drzewa, stany szkieletu, zmiany szkieletu na bazie linearyzacji drzewa acclaim - dane pomocnicze przy wyznaczaniu danych ruchu, identyfikator kości biovision - dane ruchu jako double imucostumelib - częstotliwość odświeżania kostiumu częścią jego stanu, inicjalizacja przy tworzeniu wielowątkowy ScopeTimerPrinter dodatnie ArrayTraits
Poprawki core:
Porządki w Filesystem - usunięcie sygnatur z std::string - wymagany jest Path
CoreAction - dodanie Q_OBJECT
plugin_imuCostume/plugin_kinematic - poprawki wynikające ze zmian w kinematiclib (używanie metod i struktur linearyzujących strukturę drzewa)


Ogólne poprawki:
usunięcie move z return - kompilator ma dowolność przy optymalizacji RVO
używanie list inicjalizacyjnych, w szczególności przy wstawianiu do różnych kolekcji
dodatnie utilities - decltype i declval dla template

git-svn-id: https://localhost:8443/svn/hm-mde-1@1303 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edrutils/include/biovisionformatslib/MotionData.h
@@ -14,11 +14,13 @@ namespace biovision
 {
 	struct MotionData
 	{
-		typedef std::vector<float> ChannelData;
+		typedef std::vector<double> ChannelData;
 		typedef std::vector<ChannelData> FrameJointData;
 		typedef std::vector<FrameJointData> Frames;
 
+		//! Czas pojedynczej ramki
 		float frameTime;
+		//! Ramki danych
 		Frames frames;
 	};
 }
